crHelloWorld/requirements.txt
functions-framework==3.*
google-auth
google-cloud-logging
python-dotenv